% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApexChart.R
\name{circleGauge}
\alias{circleGauge}
\title{Create a Circle Gauge}
\usage{
circleGauge(id, value)
}
\arguments{
\item{id}{The ID of the gauge element}

\item{value}{The value to display in the gauge (numeric)}
}
\value{
An HTML element containing the circle gauge
}
\description{
Create a Circle Gauge
}
