% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApexChart.R
\name{donutChart}
\alias{donutChart}
\title{Create a Donut Chart}
\usage{
donutChart(id, df, data_labels_enabled = TRUE, chart_title = "")
}
\arguments{
\item{id}{The ID of the chart element}

\item{df}{A data frame with two columns: labels and values}

\item{data_labels_enabled}{A boolean to enable or disable data labels (default: TRUE)}

\item{chart_title}{The title of the chart (default: "")}
}
\value{
An HTML element containing the donut chart
}
\description{
Create a Donut Chart
}
