% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApexChart.R
\name{radialChart}
\alias{radialChart}
\title{Create a Radial Bar Chart}
\usage{
radialChart(id, df, color = "#0259E6")
}
\arguments{
\item{id}{The ID of the chart element}

\item{df}{A data frame with two columns: labels and values}

\item{color}{A hexadecimal color code (default: "#0259e6")}
}
\value{
An HTML element containing the radial bar chart
}
\description{
Create a Radial Bar Chart
}
